Complete rework of adding track, consolidated it with adding driver, renamed files

+++ RacedayStopwatch/Add Item/AddItemViewController.swift
@@ -9,7 +9,7 @@
 import UIKit
 
 class AddItemViewController: UIViewController {
-
+    
     @IBOutlet weak var popupView: UIView!
     @IBOutlet weak var addItemLabel: UILabel!
     @IBOutlet weak var itemNameLabel: UILabel!
@@ -27,18 +27,20 @@ class AddItemViewController: UIViewController {
     @IBOutlet weak var addItemButton: UIButton!
     
     var driver: Driver?
+    var itemIsDriver = false
     var driverImage: UIImage?
     var helmetImage: UIImage?
     
     var track: Track?
+    var itemIsTrack = false
     var trackImage: UIImage?
     
     override func viewDidLoad() {
         super.viewDidLoad()
         popupView.layer.cornerRadius            = Constants.cornerRadius
         popupView.layer.masksToBounds           = true
-        addItemButton.layer.cornerRadius      = Constants.cornerRadius
-        addItemButton.layer.masksToBounds     = true
+        addItemButton.layer.cornerRadius        = Constants.cornerRadius
+        addItemButton.layer.masksToBounds       = true
         cancelButton.layer.cornerRadius         = Constants.cornerRadius
         cancelButton.layer.masksToBounds        = true
         
@@ -48,20 +50,46 @@ class AddItemViewController: UIViewController {
         imageContainerView.isHidden            = true
         
         pictureOrImageControl.addTarget(self, action: #selector(changeContainerView), for: .valueChanged)
-        addItemLabel.text             = Constants.ADD_DRIVER_LABEL
-        itemNameLabel.text            = Constants.DRIVER_NAME_PLACEHOLDER
-        itemNumberLabel.text          = Constants.DRIVER_NUMBER_PLACEHOLDER
+        
+        
         addDoneButton()
         pictureOrImageLabel.text       = Constants.ADD_DRIVER_PICTURE_OR_HELMET_LABEL
         pictureOrImageLabel.isHidden   = true
-        pictureOrImageControl.setTitle(Constants.ADD_DRIVER_PICTURE_SEGMENT, forSegmentAt: 0)
-        pictureOrImageControl.setTitle(Constants.ADD_DRIVER_HELMET_SEGMENT, forSegmentAt: 1)
-        if let driver = driver {
-            itemName.text     = driver.name
-            itemNumber.text   = driver.number
-            
-            addItemLabel.text = Constants.EDIT_DRIVER_LABEL
-            addItemButton.setTitle(Constants.SAVE_BUTTON_TITLE, for: .normal)
+        
+        pictureOrImageControl.setTitle(Constants.ADD_ITEM_PICTURE_SEGMENT, forSegmentAt: 0)
+        addItemButton.setTitle(Constants.SAVE_BUTTON_TITLE, for: .normal)
+        itemNameLabel.text   = Constants.ITEM_NAME_LABEL
+
+        if itemIsDriver{
+            pictureOrImageControl.setTitle(Constants.ADD_ITEM_HELMETS_SEGMENT, forSegmentAt: 1)
+            itemNumberLabel.text = Constants.DRIVER_NUMBER_LABEL
+            if let driver = driver {
+                print("EDITING EXISTING DRIVER")
+                itemName.text     = driver.name
+                itemNumber.text   = driver.number
+                
+                addItemLabel.text = Constants.EDIT_DRIVER_LABEL
+            }else{
+                print("NEW DRIVER SOM ER EMPTY")
+                addItemLabel.text        = Constants.ADD_DRIVER_LABEL
+                itemName.placeholder     = Constants.DRIVER_NAME_PLACEHOLDER
+                itemNumber.placeholder   = Constants.DRIVER_NUMBER_PLACEHOLDER
+            }
+        }else if itemIsTrack{
+            pictureOrImageControl.setTitle(Constants.ADD_ITEM_TRACKS_SEGMENT, forSegmentAt: 1)
+            itemNumberLabel.text = Constants.TRACK_LENGTH_LABEL
+            if let track = track {
+                print("EDITING EXISTING TRACK")
+                itemName.text     = track.name
+                itemNumber.text   = String(track.length)
+                
+                addItemLabel.text = Constants.EDIT_TRACK_LABEL
+            }else{
+                print("NEW TRACK SOM ER EMPTY")
+                addItemLabel.text             = Constants.ADD_TRACK_LABEL
+                itemName.placeholder          = Constants.TRACK_NAME_PLACEHOLDER
+                itemNumber.placeholder        = Constants.TRACK_LENGTH_PLACEHOLDER
+            }
         }
     }
     //refactor to an extension as it's being used multiple places
@@ -75,11 +103,11 @@ class AddItemViewController: UIViewController {
     }
     
     override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
-        if (segue.identifier == "embedHelmetPickerView"){
-            let selectHelmet = segue.destination as! HelmetPickerViewController
-            selectHelmet.delegate   = self
+        if (segue.identifier == "embedImagePickerView"){
+            let selectItem = segue.destination as! ImagePickerViewController
+            selectItem.delegate   = self
         }else if (segue.identifier == "embedPicturePickerView"){
-            let selectPicture = segue.destination as! DriverPictureViewController
+            let selectPicture = segue.destination as! ItemPictureViewController
             selectPicture.delegate  = self
             selectPicture.driver    = driver
         }
@@ -119,24 +147,32 @@ class AddItemViewController: UIViewController {
         }
         
         //edit an existing driver
-        if let driver = driver{
-            driver.name     = name
-            driver.number   = number
-            driver.image    = image.pngData()
-            CoreDataService.saveContext()
-            dismiss(animated: true, completion: nil)
-        }
-        //save a new driver
-        else {
-            let driver = Driver(context: CoreDataService.context)
-            driver.name     = name
-            driver.number   = number
-            driver.image    = image.pngData()
-            if (driver.name!.count > 0 && driver.number!.count > 0){
+        if itemIsDriver{
+            if let driver = driver{
+                driver.name     = name
+                driver.number   = number
+                driver.image    = image.pngData()
                 CoreDataService.saveContext()
                 dismiss(animated: true, completion: nil)
-            }else{
-                CoreDataService.context.delete(driver)
+            }
+            //save a new driver
+            else {
+                let driver = Driver(context: CoreDataService.context)
+                driver.name     = name
+                driver.number   = number
+                driver.image    = image.pngData()
+                if (driver.name!.count > 0 && driver.number!.count > 0){
+                    CoreDataService.saveContext()
+                    dismiss(animated: true, completion: nil)
+                }else{
+                    CoreDataService.context.delete(driver)
+                }
+            }
+        }else if itemIsTrack{
+            if let track = track{
+                
+            }else{//save a new track
+                
             }
         }
     }
@@ -153,13 +189,13 @@ extension AddItemViewController: UITextFieldDelegate{
 }
 
 //TODO: - rework these to one variable and automatically understanding which one the user wants
-extension AddItemViewController: HelmetPickerProtocol{
-    func selectedHelmet(image: UIImage) {
+extension AddItemViewController: ImagePickerProtocol{
+    func selectedImage(image: UIImage) {
         helmetImage = image
     }
 }
-extension AddItemViewController: DriverPictureProtocol{
-    func selectedDriverPicture(_ image: UIImage) {
+extension AddItemViewController: ItemPictureProtocol{
+    func selectedItemPicture(_ image: UIImage) {
         driverImage = image
     }
 }
